refactor: change .jsx files into .tsx files
