--- src/App.jsx
@@ -1,11 +1,0 @@
-import { ResumeBuilder } from "./components/ResumeBuilder";
-
-function App() {
-  return (
-    <main>
-      <ResumeBuilder />
-    </main>
-  );
-}
-
-export default App;
